+++ packages/expo-image/src/ExpoImage.web.tsx
@@ -67,6 +67,7 @@ export default function ExpoImage({
   onError,
   responsivePolicy,
   onLoadEnd,
+  onDisplay,
   priority,
   blurRadius,
   recyclingKey,
@@ -131,6 +132,7 @@ export default function ExpoImage({
             onLoad: [onLoadAdapter(onLoad), onLoadEnd, onReady],
             onMount: [onMount],
             onTransitionEnd: [onAnimationFinished],
+            onDisplay: [onDisplay],
           }}
           style={{
             objectFit: selectedSource ? contentFit : imagePlaceholderContentFit,